refactor: 리더보드 build method 모두 helper 클래스로 분리
