+++ src/discord/commands/LeaderboardHelpers.tsx
@@ -1,166 +1,14 @@
-import { ApplicationCommandOptionType, AttachmentBuilder, EmbedBuilder, codeBlock } from "discord.js";
-import { Slash, SlashOption, Discord } from "discordx";
-import { StatusCodes } from "http-status-codes";
+import { EmbedBuilder, codeBlock } from "discord.js";
 import Vars from "@/Vars";
 import satori from "satori";
 import { Resvg } from "@resvg/resvg-js";
 import colors from "@radix-ui/colors";
 import { Fragment } from "react";
 import PColors from "@/constants/PColors";
-import PaginationMessageManager from "../messageManagers/PaginationMessageManager";
-import SlashOptionBuilder from "@/utils/SlashOptionBuilder";
 
-const validVersions = [
-  "cb1",
-  "closedbeta1",
-  "cb2",
-  "closedbeta2",
-  "ob",
-  "openbeta",
-  "s1",
-  "season1",
-  "s2",
-  "season2",
-  "live",
-  "s3",
-  "season3",
-  "s3worldtour",
-  "season3worldtour",
-];
-const validPlatforms = ["steam", "xbox", "psn", "crossplay"];
 const rankColor = [0xea6500, 0xd9d9d9, 0xebb259, 0xc9e3e7, 0x54ebe8, 0xe0115f];
-
-const VersionParameter = SlashOptionBuilder.create({
-  name: "버전",
-  description: "리더보드 시즌을 선택합니다. (cb1, cb2, ob, s1, s2, live, s3, s3worldtour)",
-  required: false,
-  type: ApplicationCommandOptionType.String,
-  default: "s3",
-  validators: [
-    [
-      (version) => Boolean(!version || validVersions.includes(version)),
-      `잘못된 버전입니다.
-가능한 버전 값: ${validVersions.join(", ")}`,
-    ],
-  ],
-});
-
-const PlatformParameter = SlashOptionBuilder.create({
-  name: "플랫폼",
-  description: "리더보드 플랫폼을 선택합니다. (crossplay, steam, xbox, psn)",
-  required: false,
-  type: ApplicationCommandOptionType.String,
-  default: "crossplay",
-  validators: [
-    [
-      (platform) => Boolean(!platform || validPlatforms.includes(platform)),
-      `잘못된 플랫폼입니다.
-가능한 플랫폼 값: ${validPlatforms.join(", ")}`,
-    ],
-  ],
-});
-
-@Discord()
-export default class LeaderboardService {
-  @Slash({
-    name: "랭킹목록",
-    description: "the finals의 모든 리더보드를 조회합니다.",
-  })
-  async leaderboard(
-    @SlashOption(VersionParameter)
-    version: string | undefined,
-    @SlashOption(PlatformParameter)
-    platform: string | undefined,
-    interaction: Discord.ChatInputCommandInteraction,
-  ) {
-    if (!version || !platform) return;
-
-    await interaction.deferReply();
-    const result = await fetch(`https://api.the-finals-leaderboard.com/v1/leaderboard/${version}/${platform}`)
-      .then(async (response) => ({
-        status: response.status,
-        data: (await response.json()) as LeaderboardData,
-      }))
-      .catch((e) => console.warn(e)); // print warning and ignore.
-
-    if (result === undefined || result.status != StatusCodes.OK) {
-      interaction.editReply("서버에 문제가 있어 명령어를 처리할 수 없습니다. X(");
-      return;
-    }
-    if (result.data.count === 0) {
-      interaction.editReply("검색 결과가 없습니다 (ㅠ ㅠ)");
-      return;
-    }
-
-    const manager = await PaginationMessageManager.createOnInteraction(interaction, {
-      size: ~~(result.data.count / 20),
-    });
-    const handleChange = async () => {
-      const svg = await this.buildTableImg(
-        result.data.data!.slice(manager.$currentPage * 20, (manager.$currentPage + 1) * 20),
-        platform,
-        version,
-      )!;
-      manager.messageData.files = [new AttachmentBuilder(svg)];
-      await manager.update();
-    };
-    await handleChange();
-    manager.events.on("change", handleChange);
-    manager.events.once("end", () => manager.events.off("change", handleChange));
-  }
-  @Slash({
-    name: "전적검색",
-    description: "TheFinals의 랭크를 검색합니다 (최소순위 10000위)",
-  })
-  async search(
-    @SlashOption({
-      name: "검색어",
-      description: "리더보드에서 유저를 검색합니다 (* <= 전체검색)",
-      required: true,
-      type: ApplicationCommandOptionType.String,
-    })
-    target: string,
-    @SlashOption(VersionParameter)
-    version: string | undefined,
-    @SlashOption(PlatformParameter)
-    platform: string | undefined,
-    interaction: Discord.ChatInputCommandInteraction,
-  ) {
-    if (!version || !platform) return;
-
-    await interaction.deferReply();
-    const searchTarget = target === "*" ? "" : target;
-    const result = await fetch(
-      `https://api.the-finals-leaderboard.com/v1/leaderboard/${version}/${platform}?name=${searchTarget}`,
-    )
-      .then(async (response) => ({
-        status: response.status,
-        data: (await response.json()) as LeaderboardData,
-      }))
-      .catch((e) => console.warn(e)); // print warning and ignore.
-    if (result === undefined || result.status != StatusCodes.OK) {
-      interaction.editReply("서버에 문제가 있어 전적검색을 할 수 없습니다. X(");
-      return;
-    } else if (result.data.count === 0) {
-      interaction.editReply("검색 결과가 없습니다 (ㅠ ㅠ)");
-      return;
-    }
-
-    const manager = await PaginationMessageManager.createOnInteraction(interaction, { size: result.data.count });
-    const handleChange = async () => {
-      const data = result.data.data![manager.$currentPage]; // 순서상 data가 없는건 불가능
-      const rankImgUri =
-        "league" in data ? `public/images/ranks/${data.league.toLowerCase().replaceAll(" ", "-")}.png` : "";
-
-      manager.messageData.embeds = [this.buildUserDataEmbed(data)];
-      manager.messageData.files = "league" in data ? [new AttachmentBuilder(rankImgUri)] : [];
-      await manager.update();
-      manager.events.once("change", handleChange);
-    };
-    await handleChange();
-  }
-
-  async buildTableImg(dataset: LeaderBoardUserData[], platform: string, version: string) {
+export default class LeaderboardHelpers {
+  public static async buildTableImg(dataset: LeaderBoardUserData[], platform: string, version: string) {
     const tableCells: React.JSX.Element[][] = Array.from({ length: 3 }, () => []);
     dataset.map((data) => {
       tableCells[0].push(
@@ -315,7 +163,7 @@ export default class LeaderboardService {
    *
    * @returns @typs {EmbedBuilder}
    */
-  buildUserDataEmbed(data: LeaderBoardUserData): EmbedBuilder {
+  public static buildUserDataEmbed(data: LeaderBoardUserData): EmbedBuilder {
     const builder = new EmbedBuilder()
       .setColor(PColors.primary)
       .setTitle(`${data.name}` /*`#${data.rank} - 『${data.name}』`*/)
